# #39 Refactor code to be asynchronous