--- a/src/_includes/partials/audio-player.html
+++ b/src/_includes/partials/audio-player.html
@@ -1,12 +1,15 @@
 {% set audioUrl = service %}
-
+{#
 {% set trackData = media_audio_field_track.split(':') %}
+#}
 
 <figure>
     <figcaption>Listen</figcaption>
     <audio controls>
+    {#}
         <source src="{{ ['/images/', audioUrl] | join }}" />
         <track default kind="{{ trackData[1] }}" srclang="{{ trackData[2] }}"
             src="{{ ['/images/', trackData[3]] | join }}" />
+            #}
     </audio>
 </figure>