--- a/src/islandty/ContentModels/Audio.js
+++ b/src/islandty/ContentModels/Audio.js
@@ -4,33 +4,23 @@ const defaultContentModel = require('./default.js');
 require('dotenv').config();
 
 module.exports = {
-  /**
-   * Async version of ingest
-   */
   async ingest(item, inputMediaPath, outputDir) {
     try {
-      // Run default ingestion first
       await defaultContentModel.ingest(item, inputMediaPath, outputDir);
-
-      // Process directory structure asynchronously
       await this.createDirectoryStructure(
         {
           media_audio_track: this.parseFieldTrack(
-            item['media:audio:field_track'])
-          },
-
+            item['media:audio:field_track']
+          )
+},
         inputMediaPath,
         outputDir
       );
     } catch (err) {
-      console.error(`Error in media_audio ingestion: ${err.message}`);
-      throw err;
+      throw new Error(`Media audio ingestion failed: ${err.message}`);
     }
   },
 
-  /**
-   * Sync method remains unchanged
-   */
   updateFilePaths(item) {
     defaultContentModel.updateFilePaths(item);
 
@@ -44,33 +34,38 @@ module.exports = {
   parseFieldTrack(fieldTrackStr) {
     const result = {};
 
-    // Handle empty or invalid input
     if (!fieldTrackStr || typeof fieldTrackStr !== 'string') {
-      console.warn('Invalid field track string:', fieldTrackStr);
-      return result;
+      throw new Error(`Invalid field track input: ${typeof fieldTrackStr}`);
     }
 
     const entries = fieldTrackStr.split('|').filter(entry => entry.trim() !== '');
 
-    for (const entry of entries) {
+    for (const [index, entry] of entries.entries()) {
       const parts = entry.split(':').map(part => part.trim());
 
-      // Validate minimum required parts
       if (parts.length < 4) {
-        console.warn('Skipping invalid field track entry:', entry);
-        continue;
+        throw new Error(
+          `Invalid field track entry at position ${index + 1}: ` +
+          `'${entry}' - requires at least 4 components`
+        );
       }
 
       const [category, type, lang, ...pathParts] = parts;
-      const filePath = pathParts.join(':'); // Re-join remaining parts as path
+      const filePath = pathParts.join(':');
 
-      // Validate critical components
       if (!category || !type || !lang || !filePath) {
-        console.warn('Skipping incomplete field track entry:', entry);
-        continue;
+        throw new Error(
+          `Incomplete field track entry at position ${index + 1}: ` +
+          `'${entry}' - missing required component`
+        );
       }
 
-      // Build nested structure with null checks
+      if (typeof filePath !== 'string' || filePath.trim() === '') {
+        throw new Error(
+          `Empty file path in field track entry at position ${index + 1}: '${entry}'`
+        );
+      }
+
       result[category] = result[category] || {};
       const categoryObj = result[category];
 
@@ -78,21 +73,12 @@ module.exports = {
       const typeObj = categoryObj[type];
 
       typeObj[lang] = typeObj[lang] || [];
-
-      // Validate and push path
-      if (typeof filePath === 'string' && filePath.trim() !== '') {
-        typeObj[lang].push(filePath);
-      } else {
-        console.warn('Skipping empty path in entry:', entry);
-      }
+      typeObj[lang].push(filePath);
     }
 
     return result;
   },
 
-  /**
-   * Async directory structure creation
-   */
   async createDirectoryStructure(obj, inputPath, rootPath) {
     try {
       for (const [key, value] of Object.entries(obj)) {
@@ -105,20 +91,20 @@ module.exports = {
           await fs.mkdir(currentPath, { recursive: true });
           const files = Array.isArray(value) ? value : [value];
 
-          // Add validation for file paths
-          const validFiles = files.filter(filePath =>
-            typeof filePath === 'string' && filePath.trim() !== ''
-          );
+          // Validate all files first
+          files.forEach((filePath, index) => {
+            if (typeof filePath !== 'string' || filePath.trim() === '') {
+              throw new Error(
+                `Invalid file path at position ${index + 1} in ${currentPath}: ` +
+                `'${filePath}'`
+              );
+            }
+          });
 
-          if (validFiles.length !== files.length) {
-            console.warn(`Skipped ${files.length - validFiles.length} invalid file paths in ${currentPath}`);
-          }
-
-          await Promise.all(validFiles.map(async (filePath) => {
+          await Promise.all(files.map(async (filePath) => {
             const fileName = path.basename(filePath);
             if (!fileName) {
-              console.error(`Invalid file path: ${filePath}`);
-              return;
+              throw new Error(`Empty filename in path: ${filePath}`);
             }
 
             const destPath = path.join(currentPath, fileName);
@@ -126,17 +112,16 @@ module.exports = {
 
             try {
               await fs.copyFile(sourcePath, destPath);
-              console.log(`Copied ${sourcePath} to ${destPath}`);
             } catch (err) {
-              console.error(`Failed to copy ${sourcePath} to ${destPath}:`, err);
-              throw err;
+              throw new Error(
+                `Failed to copy ${sourcePath} to ${destPath}: ${err.message}`
+              );
             }
           }));
         }
       }
     } catch (err) {
-      console.error('Error creating directory structure:', err);
-      throw err;
+      throw new Error(`Directory structure creation failed: ${err.message}`);
     }
   },