This work was done with the assistance of DeepSeek.

## Performance difference:

```shell
main
npm run production  23.46s user 22.73s system 346% cpu 13.345 total

async 
npm run production  21.97s user 19.37s system 242% cpu 17.041 total
```

The CPU usage improvement here is significant. This will also be very helpful when we start to pull assets from network sources, since one blocking request could slow down the entire process.